+++ lib/xliff/bundle.rb
@@ -57,7 +57,7 @@ module Xliff
     # @return [File, nil] The file, if found.
     def file_named(name)
       @files.find do |file|
-        file.original == name || ::File.basename(file.original) == name
+        file.original == name || (file.original && ::File.basename(file.original) == name)
       end
     end
 
@@ -70,8 +70,6 @@ module Xliff
 
       xliff_node = document.create_element('xliff')
       attach_xliff_metadata(xliff_node)
-
-      return document if @files.empty?
 
       @files.each do |file|
         xliff_node.add_child(file.to_xml)
@@ -123,13 +121,14 @@ module Xliff
     def self.from_xml(xml)
       raise if xml.nil?
 
-      raise 'Invalid XLIFF file – the root node must be `<xliff>`' if xml.document.root.name != 'xliff'
+      root = xml.document.root
+      raise 'Invalid XLIFF file – the root node must be `<xliff>`' if root.nil? || root.name != 'xliff'
 
       bundle = Bundle.new
 
-      xml.document.root.element_children
-         .select { |node| node.name == 'file' }
-         .each { |node| bundle.add_file File.from_xml(node) }
+      root.element_children
+          .select { |node| node.name == 'file' }
+          .each { |node| bundle.add_file File.from_xml(node) }
 
       bundle
     end